+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # Changelog
 
+## 0.1.2
+* Adds empty page check feature for `Dumpbuf`.
+* Adds bad block check feature for `Dumpbuf`.
+* Adds corresponding `check` subcommand.
+
 ## 0.1.1
 * Fixes read buffer clearing.
 * Sends CTRL+C before sending U-boot shell commands.